+++ tests/testthat/test-sensitivity.R
@@ -1,5 +1,6 @@
 test_that("test sensitivity analysis", {
   skip_if(Sys.getenv("CI") == "true", "only run the sensitivity tests locally")
+  skip_on_cran()
   skip_on_covr()
   set.seed(42)
   x <- rnorm(100)
@@ -67,6 +68,7 @@ test_that("test sensitivity analysis", {
 
 test_that("test sensitivity analysis (stan backend)", {
   skip_if(Sys.getenv("CI") == "true", "only run the sensitivity tests locally")
+  skip_on_cran()
   skip_on_covr()
   set.seed(42)
   x <- rnorm(100)