+++ tests/testthat/test-traceplots.R
@@ -60,6 +60,7 @@ test_that("test traceplots for the pre-model", {
 
 test_that("test traceplots for the multivariate model (stan backend)", {
   skip_if(Sys.getenv("CI") == "true", "only run the Stan tests locally")
+  skip_on_cran()
   skip_on_covr()
   set.seed(42)
   x <- rnorm(100)
@@ -100,6 +101,7 @@ test_that("test traceplots for the multivariate model (stan backend)", {
 
 test_that("test traceplots for the pre-model (stan backend)", {
   skip_if(Sys.getenv("CI") == "true", "only run the Stan tests locally")
+  skip_on_cran()
   skip_on_covr()
   set.seed(42)
   x <- rnorm(100)