+++ NEWS.md
@@ -1,5 +1,9 @@
-# rACMEMEEV (development version)
+# rACMEMEEV 1.0.1
+
+### Bug Fixes
+
+* Fixing error on Fedora GCC sensitivity analysis test
+
+## rACMEMEEV 1.0.0
 
 * Initial CRAN submission.
-
-# rACMEMEEV 1.0.0